requirements.txt
requests
cloudinary
pillow
python-dotenv
colorama
python-dotenv
fake-useragent